Translation units for data struct and delimeters
