+++ namestnikov.kirill/T2/delimeters.cpp
@@ -1,0 +1,57 @@
+#include "delimeters.hpp"
+
+std::istream & namestnikov::operator>>(std::istream & in, DelimeterChar && del)
+{
+  std::istream::sentry guard(in);
+  if (!guard)
+  {
+    return in;
+  }
+  char c = 0;
+  in >> c;
+  if (c != del.delimeter)
+  {
+    in.setstate(std::ios::failbit);
+  }
+  return in;
+}
+
+std::istream & namestnikov::operator>>(std::istream & in, DelimeterLetter && del)
+{
+  std::istream::sentry guard(in);
+  if (!guard)
+  {
+    return in;
+  }
+  char c = 0;
+  in >> c;
+  c = std::tolower(c);
+  if (c != del.delimeter)
+  {
+    in.setstate(std::ios::failbit);
+  }
+  return in;
+}
+
+std::istream & namestnikov::operator>>(std::istream & in, DelimeterString && del)
+{
+  std::istream::sentry guard(in);
+  if (!guard)
+  {
+    return in;
+  }
+  size_t i = 0;
+  while (del.delimeter[i] != '\0')
+  {
+    if (std::isalpha(del.delimeter[i]))
+    {
+      in >> DelimeterLetter{del.delimeter[i]};
+    }
+    else
+    {
+      in >> DelimeterChar{del.delimeter[i]};
+    }
+    ++i;
+  }
+  return in;
+}
